# docs/svc/openapi.yaml
components:
    schemas:
...
        DependencyListItem:
            properties:
                dependency_type:
                    type: string
                ecosystem:
                    type: string
                id:
                    format: int32
                    type: integer
                identifier:
                    type: string
                version:
                    type: string
                version_constraint:
                    type: string
            type: object
        DependencyListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/DependencyListItem'
                    type: array
            type: object
...
        ErrorResponse:
            properties:
                error:
                    type: string
            required:
                - error
            type: object
        GetProjectResponse:
            properties:
                created_at:
                    type: string
                id:
                    format: int32
                    type: integer
                name:
                    type: string
                source_type:
                    type: string
                updated_at:
                    type: string
            type: object
...
        ProjectListItem:
            properties:
                created_at:
                    type: string
                id:
                    format: int32
                    type: integer
                name:
                    type: string
                source_type:
                    type: string
                updated_at:
                    type: string
            type: object
        ProjectListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/ProjectListItem'
                    type: array
            type: object
        ProjectSummaryResponse:
            properties:
                project_id:
                    type: integer
                summary:
                    items:
                        $ref: '#/components/schemas/SummaryRow'
                    type: array
            type: object
...
        SummaryRow:
            properties:
                behavior_passed:
                    format: int32
                    type: integer
                dependency_type:
                    type: string
                has_attestation:
                    format: int32
                    type: integer
                has_oss_rebuild:
                    format: int32
                    type: integer
                total:
                    format: int32
                    type: integer
            type: object
...
        UploadProjectRequest:
            properties:
                file:
                    type: string
                name:
                    type: string
            type: object
        UploadProjectResponse:
            properties:
                direct_deps:
                    type: integer
                id:
                    format: int32
                    type: integer
                name:
                    type: string
                source_type:
                    type: string
                total_deps:
                    type: integer
            type: object
info:
    description: API for querying package information and metadata
    title: Secure Package Registry API
    version: 1.0.0
openapi: 3.0.3
paths:
...
    /api/v1/projects:
        get:
            description: List all projects belonging to the authenticated user.
            operationId: listProjects
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProjectListResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: List projects
            tags:
                - Projects
        post:
            description: Create or replace a project's dependency set from a lockfile or package manifest.
            operationId: uploadProject
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UploadProjectRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UploadProjectResponse'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Upload project
            tags:
                - Projects
    /api/v1/projects/{projectID}:
        delete:
            description: Delete a project and all its dependencies.
            operationId: deleteProject
            parameters:
                - in: path
                  name: projectID
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Delete project
            tags:
                - Projects
        get:
            description: Get a single project by ID.
            operationId: getProject
            parameters:
                - in: path
                  name: projectID
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetProjectResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get project
            tags:
                - Projects
    /api/v1/projects/{projectID}/dependencies:
        get:
            description: List dependencies for a project, optionally filtered by type.
            operationId: listDependencies
            parameters:
                - in: path
                  name: projectID
                  required: true
                  schema:
                    type: string
                - description: 'Filter by dependency type: ''direct'' or ''transitive'''
                  in: query
                  name: type
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DependencyListResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: List project dependencies
            tags:
                - Projects
    /api/v1/projects/{projectID}/summary:
        get:
            description: Get aggregated security posture statistics for a project.
            operationId: getProjectSummary
            parameters:
                - in: path
                  name: projectID
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProjectSummaryResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get project summary
            tags:
                - Projects
...